refactor(random): rename to drop redundant prefixes
Two file renames following review:

- `packages/random/random.types.d.ts` -> `types.d.ts` (the qualified
  prefix stutters with the package directory; switch the consumer
  sites in `src/{int,random,uint}.js` to the `@import` tag form
  while the import paths are touched).
- `packages/random/read-uint.js` and `src/read-uint.js` -> `uint.js`
  (the file already lives under `random/src/`; the `read-` prefix is
  redundant).

Also delete the `packages/random/test/index.test.js` placeholder
stub.
