--- packages/random/test/index.test.js
@@ -1,5 +1,0 @@
-import test from '@endo/ses-ava/prepare-endo.js';
-
-test.skip('placeholder', async t => {
-  t.fail('TODO: add tests');
-});
