+++ packages/random/uint.js
@@ -4,4 +4,4 @@ export {
   randomUint24,
   randomUint32,
   randomUint53,
-} from './src/read-uint.js';
+} from './src/uint.js';
